Don't need santizeUrl anymore
